Finished terminal
-Terminal prints on its own thread
-Generates a random obstacle on the map
-The player can now move based on the accelerometer values

+++ hal/include/hal/accelerometer.h
@@ -15,9 +15,9 @@ class Accelerometer{
         int getZ();
         void updateAccValues();
     private:
-        float x;
-        float y;
-        float z;
+        int16_t x;
+        int16_t y;
+        int16_t z;
         int i2cFileDesc;
         unsigned char regAddr;
 };
